requirements.txt
groq
python-dotenv
rich
typer